+++ Demo/Keyboard/KeyboardViewController.swift
@@ -82,14 +82,10 @@ class KeyboardViewController: KeyboardInputViewController {
         // keyboardAppearance = <Insert your own custom appearance>
         // view.backgroundColor = UIColor(keyboardAppearance.keyboardBackgroundColor)
         
-        // Setup the extension to use the keyboardView below
-        // 💡 This is overwritten if Pro is registered below
-        // setup(with: keyboardView)
-
-        // Setup KeyboardKit Pro. This unlocks more features.
-        // 💡 Disable this line to run demo without Pro mode.
-        //
-        setupPro(withLicenseKey: "299B33C6-061C-4285-8189-90525BCAF098", view: keyboardView)
+        // Setup the extension to use the keyboardView below,
+        // either without or with Pro enabled.
+        setup(with: keyboardView)
+        // setupPro(withLicenseKey: "299B33C6-061C-4285-8189-90525BCAF098", view: keyboardView)
     }